mcp/mcp-obs/pyproject.toml
[project]
name = "mcp-obs"
version = "0.1.0"
description = "MCP server for observability (VictoriaLogs and VictoriaTraces)"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.26.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_obs"]
